bot/requirements.txt
aiofiles==24.1.0
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
dotenv==0.9.9
frozenlist==1.5.0
grpcio==1.71.0
grpcio-tools==1.71.0
idna==3.10
logging==0.4.9.6
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
protobuf==5.29.4
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.19.0
